model types [nfc]: Move UserStatus to modelTypes.js
